backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
llama-cpp-python
pydantic
python-dotenv
mysql-connector-python
bcrypt
aiofiles
